# jobs: add job_id index to crdb_internal.jobs virtual table 