--- a/pkg/bench/rttanalysis/BUILD.bazel
+++ b/pkg/bench/rttanalysis/BUILD.bazel
@@ -43,6 +43,7 @@ go_test(
         "generate_objects_bench_test.go",
         "grant_revoke_bench_test.go",
         "grant_revoke_role_bench_test.go",
+        "jobs_test.go",
         "orm_queries_bench_test.go",
         "rtt_analysis_test.go",
         "system_bench_test.go",
@@ -60,12 +61,15 @@ go_test(
     shard_count = 16,
     deps = [
         "//pkg/base",
+        "//pkg/jobs/jobspb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
+        "//pkg/security/username",
         "//pkg/server",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
+        "//pkg/util/protoutil",
         "//pkg/util/randutil",
     ],
 )