--- a/pkg/sql/delegate/show_changefeed_jobs.go
+++ b/pkg/sql/delegate/show_changefeed_jobs.go
@@ -23,59 +23,53 @@ func (d *delegator) delegateShowChangefeedJobs(n *tree.ShowChangefeedJobs) (tree
 	// Note: changefeed_details may contain sensitive credentials in sink_uri. This information is redacted when marshaling
 	// to JSON in ChangefeedDetails.MarshalJSONPB.
 	const (
-		// In 23.1, we can use the job_type column to filter jobs.
-		queryTarget23_1 = `
-			crdb_internal.system_jobs
-  			WHERE job_type = 'CHANGEFEED'
-		`
 		baseSelectClause = `
 WITH payload AS (
-  SELECT 
-    id, 
+  SELECT
+    id,
     crdb_internal.pb_to_json(
-      'cockroach.sql.jobs.jobspb.Payload', 
+      'cockroach.sql.jobs.jobspb.Payload',
       payload, false, true
-    )->'changefeed' AS changefeed_details 
-  FROM 
-    %s
-) 
-SELECT 
-  job_id, 
-  description, 
-  user_name, 
-  status, 
-  running_status, 
-  created, 
-  started, 
-  finished, 
-  modified, 
-  high_water_timestamp, 
-  error, 
+    )->'changefeed' AS changefeed_details
+  FROM
+  crdb_internal.system_jobs
+  WHERE job_type = 'CHANGEFEED'%s
+)
+SELECT
+  job_id,
+  description,
+  user_name,
+  status,
+  running_status,
+  created,
+  started,
+  finished,
+  modified,
+  high_water_timestamp,
+  error,
   replace(
-    changefeed_details->>'sink_uri', 
+    changefeed_details->>'sink_uri',
     '\u0026', '&'
-  ) AS sink_uri, 
+  ) AS sink_uri,
   ARRAY (
-    SELECT 
+    SELECT
       concat(
-        database_name, '.', schema_name, '.', 
+        database_name, '.', schema_name, '.',
         name
-      ) 
-    FROM 
-      crdb_internal.tables 
-    WHERE 
+      )
+    FROM
+      crdb_internal.tables
+    WHERE
       table_id = ANY (descriptor_ids)
-  ) AS full_table_names, 
+  ) AS full_table_names,
   changefeed_details->'opts'->>'topics' AS topics,
-  COALESCE(changefeed_details->'opts'->>'format','json') AS format 
-FROM 
-  crdb_internal.jobs 
+  COALESCE(changefeed_details->'opts'->>'format','json') AS format
+FROM
+  crdb_internal.jobs
   INNER JOIN payload ON id = job_id`
 	)
 
-	selectClause := fmt.Sprintf(baseSelectClause, queryTarget23_1)
-
-	var whereClause, orderbyClause string
+	var whereClause, innerWhereClause, orderbyClause string
 	if n.Jobs == nil {
 		// The query intends to present:
 		// - first all the running jobs sorted in order of start time,
@@ -87,8 +81,10 @@ FROM
 	} else {
 		// Limit the jobs displayed to the select statement in n.Jobs.
 		whereClause = fmt.Sprintf(`WHERE job_id in (%s)`, n.Jobs.String())
+		innerWhereClause = fmt.Sprintf(` AND id in (%s)`, n.Jobs.String())
 	}
 
+	selectClause := fmt.Sprintf(baseSelectClause, innerWhereClause)
 	sqlStmt := fmt.Sprintf("%s %s %s", selectClause, whereClause, orderbyClause)
 
 	return d.parse(sqlStmt)