To produce rows, `crdb_internal.jobs` queries
`crdb_internal.system_jobs`. The latter is very expensive in the
abscence of a filter on job ID because it does 2 full scans of the
job_info table.

This is a backportable change that adds a job_id index to the
crdb_internal.jobs virtual table, allowing it to produce a query to
crdb_internal.system_jobs that takes advantage of system_job's index
on id.

We should pursue a more complete fix, perhaps by changing these
virtual tables to virtual views, allowing the optimizer to solve many
of these problems for us. For example, this change does not
substantially improve `SHOW CHANGEFEED JOB <id>` without a follow up
commit.

Before this change, `SHOW JOB 101` on a fresh demo cluster produced
the following scans:

```
Scan /Tenant/2/Table/15/1/10{1-2}, /Tenant/2/Table/15/1/10{2-3}, /Tenant/2/Table/15/1/10{3-4}, /Tenant/2/Table/15/1/10{4-5}, /Tenant/2/Table/15/1/9397966460656353{29-30}, /Tenant/2/Table/15/1/93979664607451545{7-8}, /Tenant/2/Table/15/1/93979664613556224{1-2}, /Tenant/2/Table/15/1/93979664625333043{3-4}, /Tenant/2/Table/15/1/93979664629252096{1-2}, /Tenant/2/Table/15/1/93979664633554534{5-6}, /Tenant/2/Table/15/1/93979664639180800{1-2}, /Tenant/2/Table/15/1/93979664643502899{3-4}, /Tenant/2/Table/15/1/93979664647874150{5-6}, /Tenant/2/Table/15/1/93979664652251955{3-4}, /Tenant/2/Table/15/1/93979664656652697{7-8}, /Tenant/2/Table/15/1/9397966468795924{49-50}, /Tenant/2/Table/15/1/93979664688050995{3-4}, /Tenant/2/Table/15/1/93979664688146022{5-6}, /Tenant/2/Table/15/1/93979664694434201{7-8}, /Tenant/2/Table/15/1/93979664816006758{5-6}, /Tenant/2/Table/15/1/93979664825014681{7-8}, /Tenant/2/Table/15/1/93979664834094694{5-6}, /Tenant/2/Table/15/1/93979664842948608{1-2}, /Tenant/2/Table/15/1/93979664852251443{3-4}
Scan /Tenant/2/Table/53/{1-2}
Scan /Tenant/2/Table/53/{1-2}
```

Afer, it produces

```
Scan /Tenant/2/Table/15/1/10{1-2}
Scan /Tenant/2/Table/53/1/101/"legacy_payload"{-/PrefixEnd}
Scan /Tenant/2/Table/53/1/101/"legacy_progress"{-/PrefixEnd}
```

Release note: None

Epic: None