@rafiss 2024-02-02
IMO, using a virtual view would be an even less risky change. if we already know how to convert to that, i would advocate for doing it and skipping this virtual index work.