@stevendanna 2024-02-05
@rafiss I agree we should just get on with moving this to a virtual table and have started that work, but still think this might be the better first step to backport because I'd like to be able to backport this pretty far, including to versions where we have much more version-conditioned logic on what query is running.  Changing crdb_internal.system_jobs to a view is trivial probably trivial.  `crdb_internal.jobs` will be a little more involved but shouldn't be hard on master, but might be a bit more complicated on older versions.