@stevendanna 2024-02-15
@rafiss Thanks for the nudge. Unfortunately it appears that without a large number of jobs, the rtt analysis package really doesn't capture the difference.  I create a benchmark in the jobs package that creates a large number of jobs and found that this does improve things pretty substantially.  If you feel strongly about having this in the rttanalysis package, but creating the jobs is pretty slow and I didn't want to slow down a test package I don't look at often.

```
> benchstat old.txt new.txt 
goos: darwin
goarch: arm64
                    │    old.txt    │               new.txt                │
                    │    sec/op     │    sec/op     vs base                │
ShowJob/SHOW_JOB-10   185.658m ± 4%   6.306m ± 13%  -96.60% (p=0.000 n=10)
```