@rafiss 2024-02-15
i do think it would be nice to keep the rttanalysis test to make sure there's not a regression in roundtrips for SHOW JOBS.

the framework lets you do some setup. for example, to test this liquibase query, we have this step to [create 15 databases with 40 tables each](https://github.com/cockroachdb/cockroach/blob/cc6ca026319024800395293b0fb18f05dd8eb50e/pkg/bench/rttanalysis/orm_queries_bench_test.go#L21), which is executed [here](https://github.com/cockroachdb/cockroach/blob/cc6ca026319024800395293b0fb18f05dd8eb50e/pkg/bench/rttanalysis/orm_queries_bench_test.go#L772).

for testing jobs, would it work to insert fake jobs rows into `system.jobs/job_info` directly? (might need something so the setup can do more privileged operations like that. if you find yourself fighting the framework too much, then no need to sink time into it. i might pick it up on my own after you merge this.)