@stevendanna 2024-02-16
Added an rttanalysis test.  In this output the "new" code is the "base" so this shows how much worse the old thing is rather than how much better the new thing is:

```
                 │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a5840326441bb3f715aaea3dbb9c68ab9b │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a584
0326441bb3f715aaea3dbb9c68ab9b^ │                                                                                                                                                                                   
                 │                                                     sec/op                                                     │                                        sec/op                                   
       vs base                  │                                                                                                                                                                                   
Jobs/show_job-10                                                                                                     12.58m ± 22%                                                                          210.18m ±
 107%  +1571.45% (p=0.000 n=10)                                                                                                                                                                                     
                                                                                                                                                                                                                    
                 │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a5840326441bb3f715aaea3dbb9c68ab9b │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a584
0326441bb3f715aaea3dbb9c68ab9b^ │                                                                                                                                                                                   
                 │                                                   roundtrips                                                   │                                       roundtrips                                
        vs base                 │                                                                                                                                                                                   
Jobs/show_job-10                                                                                                       3.000 ± 0%                                                                               8.00
0 ± 0%  +166.67% (p=0.000 n=10)                                                                                                                                                                                     
                                                                                                                                                                                                                    
                 │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a5840326441bb3f715aaea3dbb9c68ab9b │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a584
0326441bb3f715aaea3dbb9c68ab9b^ │                                                                                                                                                                                   
                 │                                                      B/op                                                      │                                          B/op                                   
        vs base                 │                                                                                                                                                                                   
Jobs/show_job-10                                                                                                    18.55Mi ± 34%                                                                             73.12M
i ± 5%  +294.14% (p=0.000 n=10)

                 │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a5840326441bb3f715aaea3dbb9c68ab9b │ /var/folders/_n/gm6dl7_d7sx1zzkqc4h18l0h0000gn/T/tmp.jNYQCstRma/bench.c46938a584
0326441bb3f715aaea3dbb9c68ab9b^ │
                 │                                                   allocs/op                                                    │                                       allocs/op                                 
       vs base                  │
Jobs/show_job-10                                                                                                     56.03k ± 11%                                                                           1114.71k
 ± 7%  +1889.35% (p=0.000 n=10)
```